requirements.txt
streamlit==1.41.1
numpy==1.23.5
scikit-learn==1.6.0
scipy==1.10.1
pandas==2.2.3
python-dotenv==1.0.1
langchain-core==0.3.29
langchain-groq==0.2.3
langdetect==1.0.9
langsmith==0.2.10
textract
regex==2024.11.6
requests==2.32.3
gensim==3.8.3
beautifulsoup4
selenium==4.27.1
urllib3==2.3.0